_48_rotate

+++ src/main/java/LeeCode/_50_MyPow.java
@@ -36,14 +36,14 @@ public class _50_MyPow {
     }
 
     public static double pow3(double x, int n) {
-        int N = n;
+        long N = n;
         if (N < 0) {
             x = 1 / x;
             N = -N;
         }
         double ans = 1;
         double current_product = x;
-        for (int i = N; i > 0; i /= 2) {
+        for (long i = N; i > 0; i /= 2) {
             if ((i % 2) == 1)
                 ans = ans * current_product;
             current_product = current_product * current_product;
